backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
numpy<2
ollama==0.4.8
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
protobuf==5.29.4
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.2
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
